requirements/_base.txt
asgiref==3.2.10
Django==3.0.8
djangorestframework==3.11.0
pkg-resources==0.0.0
psycopg2-binary==2.8.5
pytz==2019.2
sqlparse==0.3.0
paramiko==2.7.1
django-cors-headers==3.4.0
django-polymorphic==2.1.2